Remove `Gridding`
