+++ src/test/scala/vectorpipe/osm/internal/Pow2LayoutSpec.scala
@@ -113,4 +113,14 @@ class Pow2LayoutSpec extends FunSpec with Matchers {
     }
 
   }
+
+  describe("Powers of 2") {
+    it("isPow2") {
+      (1 to 10).map(n => Pow2Layout.isPow2(n)) shouldBe Seq(true, true, false, true, false, false, false, true, false, false)
+    }
+
+    it("pow2Ceil") {
+      (1 to 10).map(n => Pow2Layout.pow2Ceil(n)) shouldBe Seq(1, 2, 4, 4, 8, 8, 8, 8, 16, 16)
+    }
+  }
 }
